Remove specific mention of Minecraft
I've thought long and hard about this, and I don't think it's necessary to specifically mention Minecraft. The intent is that this license will be used to license Minecraft Mods, but @asiekierka made the point that some mods want to target multiple platforms. This possibility was brought home to me when @Soulrift asked me about the possibility of porting Minechem to TUG.

I don't think it's a perfect solution, and I need to re-read the text of the license to make sure this doesn't open a bunch of loopholes. Also, I'll need to come up with a name other than MMPL, as the first M is "Minecraft", which may no longer apply.